Split mahjax typesetter into lib package
