+++ dev_mode/package.json
@@ -50,6 +50,7 @@
     "@jupyterlab/mainmenu": "^0.7.4",
     "@jupyterlab/mainmenu-extension": "^0.7.5",
     "@jupyterlab/markdownviewer-extension": "^0.18.4",
+    "@jupyterlab/mathjax2": "^0.6.0",
     "@jupyterlab/mathjax2-extension": "^0.6.4",
     "@jupyterlab/notebook": "^0.18.4",
     "@jupyterlab/notebook-extension": "^0.18.4",
@@ -258,6 +259,7 @@
       "@jupyterlab/mainmenu": "../packages/mainmenu",
       "@jupyterlab/mainmenu-extension": "../packages/mainmenu-extension",
       "@jupyterlab/markdownviewer-extension": "../packages/markdownviewer-extension",
+      "@jupyterlab/mathjax2": "../packages/mathjax2",
       "@jupyterlab/mathjax2-extension": "../packages/mathjax2-extension",
       "@jupyterlab/notebook": "../packages/notebook",
       "@jupyterlab/notebook-extension": "../packages/notebook-extension",
