+++ packages/mathjax2/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@jupyterlab/mathjax2-extension",
-  "version": "0.6.4",
-  "description": "A JupyterLab extension providing MathJax 2 Typesetting",
+  "name": "@jupyterlab/mathjax2",
+  "version": "0.6.0",
+  "description": "A MathJax 2 Typesetting provider for JupyterLab",
   "homepage": "https://github.com/jupyterlab/jupyterlab",
   "bugs": {
     "url": "https://github.com/jupyterlab/jupyterlab/issues"
@@ -28,16 +28,11 @@
     "watch": "tsc -w --listEmittedFiles"
   },
   "dependencies": {
-    "@jupyterlab/application": "^0.18.4",
-    "@jupyterlab/coreutils": "^2.1.4",
-    "@jupyterlab/mathjax2": "^0.6.0",
-    "@jupyterlab/rendermime": "^0.18.4"
+    "@jupyterlab/rendermime-interfaces": "^1.1.7",
+    "@phosphor/coreutils": "^1.3.0"
   },
   "devDependencies": {
     "rimraf": "~2.6.2",
     "typescript": "~2.9.2"
-  },
-  "jupyterlab": {
-    "extension": true
   }
 }
