+++ packages/mathjax2/style/index.css
