+++ packages/mathjax2/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfigbase",
+  "compilerOptions": {
+    "outDir": "./lib"
+  },
+  "include": ["src/*"]
+}
