+++ packages/metapackage/package.json
@@ -69,6 +69,7 @@
     "@jupyterlab/mainmenu": "^0.7.4",
     "@jupyterlab/mainmenu-extension": "^0.7.5",
     "@jupyterlab/markdownviewer-extension": "^0.18.4",
+    "@jupyterlab/mathjax2": "^0.6.0",
     "@jupyterlab/mathjax2-extension": "^0.6.4",
     "@jupyterlab/notebook": "^0.18.4",
     "@jupyterlab/notebook-extension": "^0.18.4",
