+++ packages/metapackage/src/index.ts
@@ -39,6 +39,7 @@ import '@jupyterlab/launcher-extension';
 import '@jupyterlab/mainmenu';
 import '@jupyterlab/mainmenu-extension';
 import '@jupyterlab/markdownviewer-extension';
+import '@jupyterlab/mathjax2';
 import '@jupyterlab/mathjax2-extension';
 import '@jupyterlab/notebook';
 import '@jupyterlab/notebook-extension';
